fix(ci build failure): adeded both plugins as dependencies
babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package without
declaring it in its dependencies.

npm warn deprecated @babel/plugin-proposal-private-property-in-object@7.21.11: This proposal has been merged to the ECMAScript standard and thus this plugin is no longer maintained. Please use @babel/plugin-transform-private-property-in-object instead.

error persisted so I uninstall the previous installed plugin, installed the above plugin and installed the previous plugin again.